apps/backend/requirements.txt
# Core dependencies
langchain
langchain-community
langchain-openai
fastapi
python-dotenv
uvicorn
pydantic
faiss-cpu
tiktoken
PyYAML>=6.0.1
numpy
openai

# Azure Core SDKs
azure-core
azure-identity
azure-cosmos>=4.5.1  # Add Cosmos DB SDK
azure-search-documents
azure-cognitiveservices-language-textanalytics
azure-ai-ml

# Monitoring
opencensus-ext-azure
opencensus-ext-requests
opencensus-ext-fastapi

# Semantic Kernel
semantic-kernel 